refactor: move swap components to swap feature folder
