+++ src/components/transactions/TxInfo/index.tsx
@@ -1,12 +1,11 @@
 import { type ReactElement } from 'react'
 import type {
-  Transfer,
-  Custom,
   Creation,
-  TransactionInfo,
+  Custom,
   MultiSend,
   SettingsChange,
-  Order,
+  TransactionInfo,
+  Transfer,
 } from '@safe-global/safe-gateway-typescript-sdk'
 import { SettingsInfoType } from '@safe-global/safe-gateway-typescript-sdk'
 import TokenAmount from '@/components/common/TokenAmount'
@@ -21,11 +20,9 @@ import {
   isSettingsChangeTxInfo,
   isTransferTxInfo,
 } from '@/utils/transaction-guards'
-import { ellipsis, formatVisualAmount, shortenAddress } from '@/utils/formatters'
+import { ellipsis, shortenAddress } from '@/utils/formatters'
 import { useCurrentChain } from '@/hooks/useChains'
-import TokenIcon from '@/components/common/TokenIcon'
-import { Box, Typography } from '@mui/material'
-import { capitalize } from '@/hooks/useMnemonicName'
+import { SwapTx } from '@/features/swap/components/SwapTxInfo/SwapTx'
 
 export const TransferTx = ({
   info,
@@ -92,49 +89,6 @@ const MultiSendTx = ({ info }: { info: MultiSend }): ReactElement => {
   )
 }
 
-const SwapTx = ({ info }: { info: Order }): ReactElement => {
-  const { kind, sellToken, sellAmount, buyToken } = info
-  const orderKindLabel = capitalize(kind)
-  const isSellOrder = kind === 'sell'
-
-  let from = (
-    <>
-      <Box style={{ paddingRight: 5, display: 'inline-block' }}>
-        <TokenIcon logoUri={sellToken.logoUri || undefined} tokenSymbol={sellToken.symbol} />
-      </Box>
-      <Typography component="span" fontWeight="bold">
-        {formatVisualAmount(sellAmount, sellToken.decimals)} {sellToken.symbol}{' '}
-      </Typography>
-    </>
-  )
-
-  let to = (
-    <>
-      <Box style={{ paddingLeft: 5, paddingRight: 5, display: 'inline-block' }}>
-        <TokenIcon logoUri={buyToken.logoUri || undefined} tokenSymbol={buyToken.symbol} />
-      </Box>{' '}
-      <Typography component="span" fontWeight="bold">
-        {buyToken.symbol}
-      </Typography>
-    </>
-  )
-  if (!isSellOrder) {
-    // switch them around for buy order
-    ;[from, to] = [to, from]
-  }
-
-  return (
-    <Box display="flex">
-      <Typography component="div" display="flex" alignItems="center" fontWeight="bold">
-        {from}
-        <Typography component="span" ml={0.5}>
-          to
-        </Typography>
-        {to}
-      </Typography>
-    </Box>
-  )
-}
 const SettingsChangeTx = ({ info }: { info: SettingsChange }): ReactElement => {
   if (
     info.settingsInfo?.type === SettingsInfoType.ENABLE_MODULE ||
